Sun, Jan 23, 2022, 3:07 AM +00:00

+++ CS50x/week1/pset1/credit/tester.c
@@ -2,18 +2,18 @@
 #include <cs50.h>
 #include <math.h>
 
-double getLengthOfLong();
+long getLengthOfLong();
 
 int main(void) {
-	double test = get_long("try me, user: ");
-	double len = getLengthOfLong(test);
+	long test = get_long("try me, user: ");
+	long len = getLengthOfLong(test);
     // long len = test%10;
-	printf("%f\n", len);
+	printf("%ld\n", len);
     return 1;
 }
 
-double getLengthOfLong(loong) {
-	double len = log10(loong);
+long getLengthOfLong(loong) {
+	long len = log10(loong);
 
     return len;
 }
